Fixed Sockets component (rgb check). Rebuild exe

+++ src/Poe/Components/Sockets.cs
@@ -43,9 +43,9 @@ namespace PoeHUD.Poe.Components
                 {
                     return list;
                 }
-                int pLinkStart = M.ReadInt(Address + 0x60);
-                int pLinkEnd = M.ReadInt(Address + 0x68);
-                int LinkGroupingCount = pLinkEnd - pLinkStart;
+                long pLinkStart = M.ReadLong(Address + 0x60);
+                long pLinkEnd = M.ReadLong(Address + 0x68);
+                long LinkGroupingCount = pLinkEnd - pLinkStart;
                 if (LinkGroupingCount <= 0 || LinkGroupingCount > 6)
                 {
                     return list;
